requirements.txt
PyYAML==6.0
numpy==1.23.5
torch==1.9.0+cuda111
